requirements.txt
transformers==4.19.2
diffusers==0.11.1
invisible_watermark==0.1.5
basicsr==1.4.2
einops==0.6.0
omegaconf==2.3.0
pytorch_lightning==1.5.9
gradio
opencv-python
pudb
imageio
imageio-ffmpeg
k-diffusion
webdataset
open-clip-torch
kornia
safetensors
timm
